requirement.txt
gradio==4.31.3
openai
ffmpeg-python
python-pptx
moviepy
python-dotenv
poppler-utils